Move MiniMax quota fetch modules into rate-limits/minimax
The five MiniMax fetch/transport modules sat flat among ~110 files covering
eight providers. Nest them so the provider's fetch surface is one directory;
credential stores (main/minimax) and the IPC handler (main/ipc) stay where
their siblings are.
