+++ src/main/ipc/minimax-credentials.ts
@@ -9,7 +9,7 @@ import {
   hasMiniMaxApiKey,
   saveMiniMaxApiKey
 } from '../minimax/minimax-api-key-store'
-import { clearMiniMaxSessionCookieJar } from '../rate-limits/minimax-request-context'
+import { clearMiniMaxSessionCookieJar } from '../rate-limits/minimax/minimax-request-context'
 import type { RateLimitService } from '../rate-limits/service'
 
 export type MiniMaxCredentialsStatus = {
