+++ src/main/rate-limits/minimax/minimax-fetcher.ts
@@ -1,5 +1,5 @@
-import type { ProviderRateLimits } from '../../shared/rate-limit-types'
-import type { MiniMaxEndpoint } from '../../shared/global-settings-types'
+import type { ProviderRateLimits } from '../../../shared/rate-limit-types'
+import type { MiniMaxEndpoint } from '../../../shared/global-settings-types'
 import {
   extractMiniMaxCookieValue,
   fetchMiniMaxWithApiKey,
